Add 'Extra' directory to repo
Note that this folder location has not been updated in the
examples which set it in the Matlab path.
Extra contains a large collection of useful Matlab functions and
scripts. Some of these are necessary for some examples to run (and
it is possible that this is also the case for the program itself).