Implements DGIGA-AFC for vector physics
DGIGA discretization remains unchanged; it represents the unlimited
case, and is common for all physics (scalar or vector PDEs).

Scalar version of AFC (as was) moved to DGIGA_AFC_scalar (low order
predictor) and AFC_scalar (anti-limiter/corrector). Scalar AFC
has been found to FAIL to provide a monotone predictor for
Burgers' (not mentioned in Kuzmin 2012).

Vector version of AFC (general implementation, valid for
both sclar and vector PDEs) is obtained by using DGIGA_AFC_vector in
combination with AFC (analogously to the scalar case). This method
circumvents the aforementioned limitation with Burgers'. Various AFC
diffusion variants are provided (tensorial and scalar). Some physics
and ICs require the most diffusive (robust) to produce a monotone
predictor. This also seems to circumvent the entropy glitch issue of
IGA (it diffuses it away), which has been found to be strongly
associated with even/odd coupling of the number of degrees of freedom
with the location of the sonic point (exact consitions that trigger it
remain to be determined). Prelimiting is done on conservative
antidiffusive fluxes, which has NOT been found to be COUNTERproductive
in Euler, and avoids typical spurious "humps" obseved in advection.

Scalar version of AFC is superseeded by its vector implementation.
Expect it to be removed in a future commit.