Applies minor fixes
Loop over edges to add (or remove) antidiffusive fluxes on an element
is now partially vectorized (loop over equations remains, but is much
cheaper).

Method that finds local extrema of each control point/node/mode in a
mesh is now a static method of AFC (which is the limiter that requires
such information).

Reflected in its implementation (via a comment) that the Rusanov Riemann
flux is also known in the literature as the Local Lax-Friedrichs flux.