Make local lax-Friedrichs the default Riemann flux (Euler)
Local lax-Friedrichs seems to be preferred in the literature. It
seems that its high diffusion does not matter in the high order
contect of DG; its efficiency, simplicity and flexibility (in
multiple dimensions) matter more.

Some authors claim that HLLC (my personal favourite until now) is
"known to be plagued by undesirable numerical phenomena in multiple
dimensions", including (but not limited to) the infamous carbuncle
issue.

So-called "robust" scalar diffusion in AFC is directly analogous to
the true Rusanov/LLF Riemann flux, while the so-called "Rusanov-like"
is not. This is reflected in the latter's implementation.