Implement TVB (not TVD) limiter of Cockburn & Shu
The TVD limiter is to be a particular case (M = 0). Not worth
implementing as a separate class.
Efficiency is not great (e.g. modminmod, repeated slope comps.).
Only works (seemingly) with the Legendre basis. Reasons unknown.

Adds an autocomplete information file (.json), which doesn't work.

Limiter information to be obtained from a method of the limiter (
e.g. to show in the solver status plots). Every limiter must
overwrite this method with its own version (abstract interface
class?).