Optimize corrected minmod operator in TVB limiter
It now looks uglier but is ~3x faster. It is now generalized to 3
2D arrays (was 3 1D column arrays, passed as a 2D array of 3
columns).

Notice that it seems to be faster to compute bult-in functions
multiple times than to store and reuse their results (WARNING:
many "aliases" I use could be adding futile overhead). Another
reason in favour of a compiled language.