Make BDF compatible with any combination of degrees
Padding in BDF now works as expected, i.e. all troubled elements
are treated as if they had the same degree, given by the maximum
degree present in the mesh (be it from a troubled element or not).
These additional Legendre modes are set to zero - i,.e. zero-padded
- which does not reduce accuracy in any way but does increase
computational costs (memory and WCT).

Reimplemented the BDF.initialize method in the process (hopefuly
it is now faster).