Fix AFC (somewhat)
AFC is now renamed to AFC_2010, foreshadowing the implementation of
other variants of the method (e.g. characteristic variables,
linearity-preserving).

Fixed a bug that caused AFC to fail completely. In doing so, the
apply method of any limiter is now divided in three, so that
initialization, stage update and time-step update are to be limited
by a different method each (in general).

AFC now has limited support for sensors. APTVD shows some promise
in detecting troubled Bspline patches (scalar advection), but
with activation at some smooth regions that remains unexplained.
The aforementioned handicap seems even more pronounced with KXRCF.