Adapt examples I to showcase input-parsing interfaces
Each example script now makes use of the novel interfaces (of
solvers and limiters).

Expect to implement a similar interface for mesh generation at some
point (should also addres unification of BCs).

Expect computation and logging of error and/or solution norms to be
moved inside the solver; should be combined with a more general
runtime state display.