Reimplement solver monitoring tasks (dedicated class)
A solver run is now monitored by an instance of the Monitor class,
designed for this purpouse. It handles runtime plotting of the
solution and norms (this is new).

Norms are now grouped in an ennumeration class, which takes care
of them in a "de facto" polymorphic way.

Only tested for Euler physics.