Implement rigorous weak-Riemann boundary conditions
Boundary conditions are enforced depending on the type of boundary
selected when instantiating a mesh. Each (numerical) boundary type
derives from a general Boundary base class. Each owns an array of
ghost elements (and NOT the mesh); these can be handles to actual
elements (e.g. in the Periodic case). Initialization and
enforcement of boundary conditions are to be made via the sealed,
public methods of the base class. These are vectorized. The array
of boundaries is, in general, heterogeneous. Boundaries are owned
by the mesh. Only tested Periodic subclass, and only for Advection
physics.

Implemented an imput parser for the Mesh class. Its constructor is
now, essentially, a full-fleged mesher (1D, structured mesh): it
incorporates the generation of edges according to various
distributions and the selection of boundary conditions.

Any simple simulation (i.e. one of the Examples I) requires two
objects, which are independent from each other: a Mesh instance
(where the approximate solution lives, and contains all information
about the discretization in space); and a Solver subclass object,
in charge of advancing any given Mesh object forward in time.

Heterogeneous bases are now supported (untested).

Polymorphic selection of Riemann solver in the Euler physics class
is now done via method name (rather than function handle), stored
at instantiation. Expected to reduce overhead (untested). Added
the centered KEP numerical flux (Jameson, 2008) as a Riemann solver
for Euler - it technically is one, isn't it? (untested).

Added a bunch of input argument checks in most constructors.

Vertical axis limits of the left Monitor subplots will increase to
keep the plotted data within bounds, but won't shrink back when the
data range reduces.