Fix the Transmissive Boundary Type
Transmissive (a.k.a. non-reflective, a.k.a. absorbing, a.k.a.
zero gradient) boundary condition now enforces solution smoothness
at the boundary. After some inverstigation, is has been found that
this is best achieved by equating each ghost Legendre coefficient
except the 1st with the bounded element; the 1st coeff. is deduced
so that the boundary value is continuous. Note that this type of
boundary employs a ghost element of degree p > 0 (same as the bound
element), which is assumed to remain constant throughout the
simulation.

Tested for Advection. Works great with TVB. Not so much with AFC,
but it seems to be the latter's fault. All seems fine.

Specifing a single Periodic boundary still seems to override the
opposite one's type. The mechanism that allows this to occur is
a mystery.

The transmissive contraint is enforced at each time-stage by
solving a sparse linear system, previously factorized into LU
triangular matrices.