Implemented Reflective with Simple Treatment
Non-simple treatment done previously failed for Euler (due to
flawed conversion to primitive variables). Simple treatment assumes
a p = 0 ghost element, is much cheaper, and avoids this problem.
However, it does NOT preserve smoothness at the boundary: it is
possible for a function and its reflected counterpart to join
smoothly (i.e. all derivatives exist and are continuous) - e.g.
cosine; the current simple treatment will, in general, only
maintain continuity in the solution (not its derivatives).

Fixed a couple of bugs related to Monitor and Norm classes.