Replaced flux control variable evaluation in DGIGA
In Fletcher's group formulation for FEA, there is no ambiguity
when evaluating flux nodal values from state nodal values.
With B-splines, it seems that treating control values as nodal
values is better (both in accuracy and efficiency).

This is a surprising result, but justifies the approach of Matthias
(which I had been using until now).

This branch should NOT be merged to master.