Deactivate failsafe limiter and override of FCT at boundaries
This changes reflect how DGIGA-AFC + FCT is currently explained in
the report.

Decided to not include the failsafe because I am pretty sure my
implementation is bugged. Cannot check because it hardly ever
activates anyway.

Decided not to override FCT at mesh edges because I suspect it could
destroy any benefit of "smart" boundary conditions, which should have
the same effect (if any) anyway. Again, this allows me to forgo any
mention of it in the report.