Hotfix: FR correction functions and projection
Found and fixed a couple of bugs in the FR implementation.

Evaluation of correction function paramer 'eta' was wrong (when eta = 0).
It now uses the order of the solution (and not the degree), as it should.

Unified L2 projection (implemented many commits back) had not been
tested for FR; I "kinda forgot" that FR does NOT use Lagrange test
functions (only trial).
Circumvented the problem quite nicely by "recycling" DGSEM's mass matrix.