Implement a batch run script for DGSEM
Added a script that computes the data for the first set of DGSEM
results. Parameters are as in the report.

Based on the temporal batch script (but streamlined substantially).
It allows an arbitrary number of files to be written, each containing
a data series to be plotted in PGFPlots. Every series is made up of an
arbitrary number of runs, in turn. The whole story about adding blank
lines has been deprecated; final output to file is now done via matlab's
tablewrite. Also, each run is now encoded as a structure (instead of a
cell array) in the parfor loop.
Added a check to prevent export of any run that did not finish sucesfully
(this removes spurious datapoints from the data files).
Lastly, output data is now entirely hardcoded (no need to preallocate
it with  NaN inputs anymore).

Fixed a bug in the adaptive quadrature algorithm (!) which could cause
it to return machine zero - apparently, it did not in most cases
purely out of luck.

Implemented a script that computes either DOFs, p or K given the
other two, for DGSEM. However, in the end, I did not use it.
Left in "Extra" just in case.