Indicate which DGIGA batches crash
Some batches refuse to be stable. I am at a loss as to why.
This seems to happen when the number of breakpoint spans is large and
the degree is > 2, and is mitigated by a large number of patches.
It only happens in Burgers.
Moreover, the issue seems to go way if Matthias' nodal DGIGA treatment
is used instead.