Hotfixes (quadrature and convergence)
Added absolute tolerance to quadvgk (10*eps) and increased the relative one
to 1e-12 (was 1e-6). Also increased the maximum number of subintervals to
10000 (was 1000), and included a warning message.
This is because I noticed a strong influence of the said tolerances in the
ErrorL2 results.

Additionally, implemented a "logarithmic distribution of integer values".
It should come in handy to distribute DOFs along the log axis when REDOING
ALL PLOTS I HAD THUS FAR!