Improve stopping criteria for quadvgk
Added some more tweaks to quadvgk; the maximum number of subintervals
is the same as MATLAB's "integral" function (same algorithm), and the
tolerances (relative and absolute) are now corrected for the
"subinterval effect" (the error in the global integral is higher than
the error in the local - subinterval-wise - one).

Decided to use rtol = 1e-9 and atol = 1e-13; both are 1000 times
smaller than MATLAB's defaults. Relative tolerance is now relative
to the global integrand.