Add MWA and Fourier footprint scripts
Implemented a batch script to generate MWA data.
It does not use the old implementation (MWA_eigen_full); instead
includes its own.

Had to redefine the gradient matrix for FR (it now is as in the report).
No idea why it was different before. Hoping that nothing breaks because
of this.

Also duplicated the Fourier footprint plotting script.
Have not decided yet if I want to use it or not; might want to
improve it if so.