Implement Fourier footprint computation inside each basis class
Implemented (in a slightly fancier way, perhaps a hair faster)
the assembly and eigendecomposition of the Fourier residual matrix
(modified wavenumber analysis) as a method in each of the basis types.
Previously, this was done by an "Extra/..." auxiliary script.

Also, now, the "Solver.optimizeCFL" method uses the newly implemented version.

A limitation of the current approach is that beta is hardcoded to 1 (upwind flux).