Speedup "Basis.getFourierFootprint" method
Found out that the vectorized approach was extremely slow for large
number of wavenumbers (e.g. when K and J are both large).
Probably the Kroenecker tensor product stuff was way over the top.

Fixed it, so that it is now reasonably fast.

Script "batch_mwa.m" now uses this method (with custom wavenumbers).