Improve the getFourierFootprint method
Managed to achieve a more robust approach to identifying which
eigenvalue correspinds to each eigenmode.
It is not thoroughly tested, but seems to work in most cases
(including FR and DGIGA's "weird" eigenmodes).

Final sorting (which mode is the physical one) is now based on
minimum L1 error with the exact dispersion relation.