Fix FR/CPR and validate the fix
I recently realized that FR was using the wrong correction function
when given either 'Ga' or 'LumpLo' options. This was caused by a wrong
use of the formulas in Vincent et al (2011).

Validated the fix by comparing correction functions and dispersion relations
with those given by Vincent et al. (2011). The former seem ok; the latter might
be questionable.

In the case of g_Ga, there seems to be a clear discrepancy with the reference
(look at the maximum value).
I can't find any bug, so I'll assume that it is they who got it wrong.