Sort such that "bubbles" are next to the physical eigenmode
Method 'Basis.getFourierFootprint' now additionally sorts the
non-physical eigenmode 'Bloch waves' according to the area that
their projection on the complex plane encloses.

This should (at least in most cases) place them first after the
physical eigenmode; i.e. eigenmode 1 is the physical one, eigenmodes
2 and 3 are the 1st bubbles (one on each side of the imaginary axis),
if any, and so on.