Add support for eigenmode-wise relative energies
Implemented the calculation of relative eigenmode energies a la Asthana
& Jameson (2015).

Also implemented a number of display methods that make it convenient to
visualize various aspects of a basis.

Verified that my 'dominant' eigenmode matches reasonably well (yet
not excellently - why?) with Alhawwary & Wang (2018), for 'central' DG.

Confirmed that my relative energies match exactly with those in Asthana
& Jameson (2015), also for FR, if I use interpolation (instead of L2 projection)
for the initial condition Fourier coefficients.