Implement combined mode analysis methods
Added a method that computes, and another that plots, the phase shift
and amplification factor obtained via combined mode analysis a la Alhawwary
and Wang, 2018.

The plotting method can also do the same for the physical, dominant (most
energized) or mean (which apparently is rubbish, by the way) modes.
Also, it works for any number of time instants.

Verified that it works against the aforementioned reference (also added a
verification test script). Only checked with DGSEM, though.