Add disp./diss ratio support (v1.1)
Added a new version of the disp./diss. ratio calculation, that now
uses the energy-weighted average ratio. It also computes the norm, but
is very slow (due to the gradient being nonsmooth as quadrature refines).

Also adpated the related display method to show, in addition to the new ratio,
the phase lag and amplification factor (by default, of the combined mode analysis).