Add export of all DGIGA candidates
The DGIGA optimization scrift can now also export a data file with the
scores (and mass matrix condition numbers) of every single DGIGA
configuration tested in search for the optimum one.

Also added the final selection of degrees in the CFL optimization
script for FR/CPR (which is still not as fancy as the others, but
I haven't really used it anyway).