Cleanup: exact solvers for Burgers
Redesigned the general exact solver for Burgers (smooth solutions only)
and moved it to the Burgers physics class.

There are now two versions of it: a fast one, and a robust one (way too
slow in practice, and not that much more robust).
Both are based on the Method Of Characteristics.

Also added a method (to the Burgers class) that automatically
estimates the breaking time (and point, and 2 of the characteristics
that intersect there) for any given initial condition.