Implement and testcustom exact solution for Burgers
The M.O.C. can be used to analytically find exact solutions to the
Burgers equation, but it falls short when they develop (or contain)
discontinuities.

I have discovered a family of initial conditions the exact solution
of which I am able to calculate (although, strictly, not prove...)
analytically for ALL times (even after the breaking time): a triangle
wave with any constant offset.

This might actually be the ONLY initial condition with FINITE non-zero
support the general exact solution of which may be analytically
obtained (at least, it does not appear in any book).