Add support for arbitary sequences of limiters
Any solver now applies as many limiters as you provide it with initially.

These are used ane ofter the other. Each is applied independently of the rest
(i.e. 'Solver' does NOT decide which elements to apply each limiter to; it
is the responsibility of each solver to do so if it neds to --- none do so,
currently, but could because I save all other limiters' activation status in
the mesh).

Checked all limiters and sensors, and fixed all trivial issues.
No guarantee that it will work 100% of the time (it is late, and I
am tired --- I probably missed something).