Add monitoring support for primal variables and additional limiters
Monitor figure now has two more toggle buttons, which cycle between
variable types and limiter priority displayed.

The limiter is not updated when clicking. I would need to store a
handle to the mesh in the monitor to do so, and I don't want to.
It gets updated next time the solver advances a stage.

Conversion to/from primitive variables is displayed at once (it does
not need the mesh). Yet, it is only possible if all 3 components are
being monitored (and if the physics are Euler, of course).