Make sensor/limiter tracking faster
Improved the implementation of the tracking of running-averaged limiter
and sensor activation by vectorizing it, and only doing it if requested
in the first place (by default, it is inactive).

'Shocks/shocks_A' now needs to explicitly requests each (primary) limiter
to store these statistics.

In an unrelated note, I now decided to use Jiang and Shu's problem
exactly as everyone else: with the 1-4-1 weights (x2), and even with
the discontinuities around the gaussian bumps on the left.