Add support for activation statistics in 'AFC_2010'
Every component of the 'isLimited' array is set to 'true' if
the difference between "uncorrected" (i.e. using *raw*
antidiffusive fluxes) and "corrected" (using limited ones) control
point values is larger than 1e-10 (absolute) or 1e-3 (relative).

'Monitor' normalizes 'isLimited' so that it is 100% when all
possible DOFs of an element are limited; this is one fewer for modal
limiters than sub-cell ones, since the former do not change mean values
(neither do the latter, but they do modify all control values).

When counting 'stats' (that "average activation probability" thing)
the distinction above is NOT made: the "prior probability" of a DOF to
be limited by a hierarchical limiter can be, at most, p/(p+1).
For AFC, this same "probability" can be 1.