Fixed some issues in batches 'C' and 'D'
The limiter stats of AFC were totally wrong whenever a sensor was used.
This is now fixed.

Batch 'C' now uses failsafes that also apply at every step (in addition
to every stage). This ensures that runs with sensors do not crash.

Changed the number of breakpoint spans in a couple of runs of batch
'D' to ensure they don't crash.
The number of DOFs is as close to the target as before (so they are still
equally well comparable).
I have no clue why they crashed before and not now; (DG)IGA is "finicky"
in this way.